pyproject.toml
[project]
name = "boardgamenights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "firebase-admin>=6.0.0",
    "python-dotenv>=1.0.0",
]

